FrameRelay_Emulation: Use Rx SeqNR == 0 to force link down
In quickly repeated test case executions we have the problem that
both the Q.933 LMI detection of link outage/interruption, as well as
the NS-ALIVE mechanism are too slow to detect the outage, and in fact
don't detect the outage.  This leads to inconsistent state on both
sides, with one side thinking NS-VCs are still ALIVE/UNBLOCKED, while
the other side considers them DEAD.

A Q.933 conforming transmittere would never send a sequece number of
zero, but always start from one.  This means we can use this as
a flag to force the peer to assume a "service affecting condition"
which will make it tell NS that the link went down, etc.

Related: OS#4974
Change-Id: I3e38ade112e11c86db906da9846a966cb33ac1bd